% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_d.R
\name{describe_d}
\alias{describe_d}
\title{Show Discriptive statistics for categorical variables}
\usage{
describe_d(.data, ...)
}
\arguments{
\item{.data}{input a tibble or data.frame}

\item{...}{categorical variables}
}
\description{
Show Discriptive statistics for categorical variables
}
