% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_d_.R
\name{describe_d_}
\alias{describe_d_}
\title{Show Discriptive statistics for categorical variables. The output is formatted with `gt` so that you can easily include it in Rmarkdown documents.}
\usage{
describe_d_(.data, ...)
}
\arguments{
\item{.data}{input a tibble or data.frame}

\item{...}{categorical variables}
}
\description{
Show Discriptive statistics for categorical variables. The output is formatted with `gt` so that you can easily include it in Rmarkdown documents.
}
