% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcomp_jack.R
\name{gcomp_jack}
\alias{gcomp_jack}
\title{Estimate the causal effect by G-computation with jackknife variance estimation.}
\usage{
gcomp_jack(
  .data,
  .outcome,
  .treatment,
  .formula_rhs,
  .estimand = c("cfmean", "ATE"),
  .weights,
  .repweights,
  .type = c("JK1", "JK2"),
  .by = NULL
)
}
\arguments{
\item{.data}{A tibble that contains the data.}

\item{.outcome}{The name of the outcome variable.}

\item{.treatment}{The name of the treatment binary variable, only allowed to be 0 or 1.}

\item{.formula_rhs}{The right-hand side of the regression formula.}

\item{.estimand}{Specify the estimand. 'cfmean' or 'ATE'.}

\item{.weights}{The name of the sampling weights.}

\item{.repweights}{The name of the replicate weights.}

\item{.type}{The variation of jackknife variance estimation. 'JK1' or 'JK2'.}

\item{.by}{The variable to be used for heterogeneity analysis. If NULL, the entire population is used.}
}
\description{
Estimate the causal effect by G-computation with jackknife variance estimation.
}
