% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haven_value_label.R
\name{haven_value_label}
\alias{haven_value_label}
\title{Create a list of value labels from SPSS or Stata data}
\usage{
haven_value_label(data)
}
\description{
Create a list of value labels from SPSS or Stata data
}
