% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haven_variable_label.R
\name{haven_variable_label}
\alias{haven_variable_label}
\title{Create a list of variable labels from SPSS or Stata data}
\usage{
haven_variable_label(data)
}
\description{
Create a list of variable labels from SPSS or Stata data
}
