% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kamaken-package.R
\docType{package}
\name{kamaken-package}
\alias{kamaken}
\alias{kamaken-package}
\title{kamaken: useful toolbox for sociologists}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Kentaro Kamada \email{kamaken0705@g.ecc.u-tokyo.ac.jp}

}
\keyword{internal}
