% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lem.R
\name{lem}
\alias{lem}
\title{Activate LEM through \code{system2()}.}
\usage{
lem(
  lat = "",
  man,
  con = "",
  dim,
  lab = "",
  mod,
  rec = "",
  des = "",
  sta = "",
  dum = "",
  dat,
  ite = "",
  see,
  path
)
}
\description{
Activate LEM through \code{system2()}.
}
