% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_cross.R
\name{my_cross}
\alias{my_cross}
\title{Cross tabulations for categorical variables}
\usage{
my_cross(.data, .x, .y, cramer = TRUE, p.value = TRUE, adjres = FALSE)
}
\arguments{
\item{.data}{input a tibble or data.frame}

\item{.x}{first variable which will be row names of the table}

\item{.y}{second variable which will be columns of the table}

\item{cramer}{if \code{TRUE}, calculates the value of cramer's V. default is \code{TRUE}}

\item{p.value}{if \code{TRUE}, calculates p value of chi-squared test. default is \code{TRUE}}

\item{adjres}{if \code{TRUE}, calculates adjusted residual and shows the results of chi-square residual tests. default is \code{FALSE}}
}
\description{
Generate a contingency table.
}
