% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_cross_plot.R
\name{my_cross_plot}
\alias{my_cross_plot}
\title{Generate a plot converted from a contingency table.}
\usage{
my_cross_plot(
  .data,
  .x,
  .y,
  strata = NULL,
  row_percent = TRUE,
  text_size = 4,
  p.value = FALSE
)
}
\arguments{
\item{.data}{input a tibble or data.frame}

\item{.x}{first variable which will be row names of the table}

\item{.y}{second variable which will be columns of the table}

\item{row_percent}{if \code{TRUE}, shows row percent with text, default is \code{TRUE}}

\item{text_size}{size of text}

\item{p.value}{if \code{TRUE}, calculates p value of chi-squared test. default is \code{TRUE}}
}
\description{
Generate a plot converted from a contingency table.
}
