% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.if.R
\name{na.if}
\alias{na.if}
\title{This is a wrapper function of \code{dplyr::na_if}. This function allows us input a numeric vector to \code{y}, while \code{dplyr::na_if} only recieves a scalar.}
\usage{
na.if(x, y)
}
\arguments{
\item{x}{a variable vector}

\item{y}{a numeric vector which contains values you want to convert to \code{NA}}
}
\description{
This is a wrapper function of \code{dplyr::na_if}. This function allows us input a numeric vector to \code{y}, while \code{dplyr::na_if} only recieves a scalar.
}
