% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_ocr.R
\name{pdf_ocr}
\alias{pdf_ocr}
\title{This function needs some outside modules tesseract, poppler and qpdf. These can be installed from homebrew.}
\usage{
pdf_ocr(path, lang = c("eng", "jpn", "jpn_vert"), dpi = 150)
}
\description{
This function needs some outside modules tesseract, poppler and qpdf. These can be installed from homebrew.
}
