% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poLCA_BLRT.R
\name{poLCA_BLRT}
\alias{poLCA_BLRT}
\title{Goodness of Fit of LCA Models with \code{poLCA_result}}
\usage{
poLCA_BLRT(
  .poLCA_result_table,
  boot_size = F,
  maxiter = 6000,
  nrep = 3,
  seed = 123,
  workers = future::availableCores()
)
}
\description{
Goodness of Fit of LCA Models with \code{poLCA_result}
}
