% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_rubin.R
\name{pool_rubin}
\alias{pool_rubin}
\title{Integration of estimation results by multiple imputation method following rubin's rule.}
\usage{
pool_rubin(.tibble, estimate = estimate, std.error = std.error, term = NULL)
}
\arguments{
\item{.tibble}{A tibble that contains the results of the estimation.}

\item{estimate}{The name of the column that contains the point estimates.}

\item{std.error}{The name of the column that contains the standard errors.}

\item{term}{When you want to integrate multiple parameters, specify the name of the columns that identifies the parameters. Allowed multiple column names like \code{c(a, b)}.}
}
\value{
A tibble that contains the integrated results.
}
\description{
Integration of estimation results by multiple imputation method following rubin's rule.
}
