% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdunif.R
\name{rdunif}
\alias{rdunif}
\title{Generate random sample from a discrete uniform distribution}
\usage{
rdunif(n, b, a = 1)
}
\description{
This function was in purrr before version 0.3.5 but deprecated in purrr 1.0.0 because it's not related to the core purpose of purrr.
}
