% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_cross_plot.R
\name{my_cross_plot}
\alias{my_cross_plot}
\title{Generate a plot converted from a contingency table.}
\usage{
my_cross_plot(
  .data,
  .x,
  .y,
  row_percent = TRUE,
  text_color = "black",
  text_size = 4,
  p.value = FALSE
)
}
\arguments{
\item{.data}{input a tibble or data.frame}

\item{.x}{first variable which will be row names of the table}

\item{.y}{second variable which will be columns of the table}

\item{row_percent}{if `TRUE`, shows row percent with text, default is `TRUE`}

\item{text_color}{color of percent text}

\item{text_size}{size of text}

\item{p.value}{if `TRUE`, calculates p value of chi-squared test. default is `TRUE`}
}
\description{
Generate a plot converted from a contingency table.
}
