% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poLCA_BLRT.R
\name{poLCA_BLRT}
\alias{poLCA_BLRT}
\title{Bootstrap Likelihood Ratio Test with `poLCA`}
\usage{
poLCA_BLRT(.poLCA_result_table, boot_size, maxiter = 6000, nrep = 3)
}
\description{
Bootstrap Likelihood Ratio Test with `poLCA`
}
