% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poLCA_check_class.R
\name{poLCA_check_class}
\alias{poLCA_check_class}
\title{format a result of latent class analysis}
\usage{
poLCA_check_class(.poLCA_result_table, nclass)
}
\description{
format a result of latent class analysis
}
