% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poLCA_result.R
\name{poLCA_result}
\alias{poLCA_result}
\title{useful short cut for \code{poLCA::poLCA}. you get a result table for reporting with the other functions like \code{poLCA_check_class} or \code{poLCA_BLRT}.}
\usage{
poLCA_result(
  formula,
  data,
  nclass,
  maxiter = 5000,
  nrep = 1,
  reorder_with = "class prop",
  reorder_outcome = 1,
  reorder_decreasing = TRUE
)
}
\description{
useful short cut for \code{poLCA::poLCA}. you get a result table for reporting with the other functions like \code{poLCA_check_class} or \code{poLCA_BLRT}.
}
